requirements.txt
absl-py==0.11.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.3
django-cors-headers==4.6.0
django-mathfilters==1.0.0
django-pwa==2.0.1
docopt==0.6.2
EbookLib==0.17.1
epub2txt==0.1.6
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.41
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
internetarchive==3.6.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2024.10.1
Levenshtein==0.27.1 
logzero==1.7.0
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
mutagen==1.47.0
mysqlclient==2.2.6
networkx==3.4.2
nltk==3.9.1
num2words==0.5.14 
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openlibrary-client @ git+https://github.com/internetarchive/openlibrary-client.git@7e45d512d031970d2bf4f56ca23f6247f205a0a8
packaging==24.2
persist-queue==1.0.0 
pillow==11.0.0
pyee==12.0.0
python-magic==0.4.27
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
rpds-py==0.22.0
schema==0.7.7
setuptools==75.4.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
uroman==1.3.1.1
python-dotenv==1.1.0